refactor(billets-app): :recycle: refactored storage
