+++ apps/billets-app/src/lib/storage/concert-detail-count-for-store-review-storage.ts
@@ -1,0 +1,11 @@
+import { mmkvKeys } from './constants'
+import { mmkvInstance } from './mmkvInstance'
+
+export const concertDetailCountForStoreReviewStorage = {
+  set: (value: number) => {
+    mmkvInstance.set(mmkvKeys.concertDetailCountForStoreReview, value)
+  },
+  get: () => {
+    return mmkvInstance.getNumber(mmkvKeys.concertDetailCountForStoreReview)
+  },
+}
