+++ apps/billets-app/src/lib/storage/constants.ts
@@ -1,4 +1,5 @@
 export const mmkvKeys = {
   currentLocation: '@billets/current-location',
   authToken: '@billets/auth-token',
+  concertDetailCountForStoreReview: '@billets/concert-detail-count-for-store-review',
 } as const
